Program to work out the smallest value of n for which the sums of the first n and n + 1 terms in the Fibonacci sequence are both divisible by 11.
Program to work out the smallest value of n for which the sums of the first n and n + 1 terms in the Fibonacci sequence are both divisible by 11. This program is coded for fixing a mathematical question raised by www.mathisfun.com.
The link is:
https://www.mathopolis.com/questions/q.php?id=1738&site=1&ref=2f6e756d626572732f6669626f6e616363692d73657175656e63652e68746d6c&title=4669626f6e616363692053657175656e6365&qs=1738_8338_8339_8340_8341_8343_3012_3071_8841_8367
The original problem is:
The partial sums of the first n and n + 1 numbers of the Fibonacci sequence are both divisible by 11.
What is the smallest value of n for which this is true?